requirements.txt
Django~=5.2.6
djangorestframework~=3.16.1
django-redis==6.0.0
celery~=5.5.3
channels==4.3.1
channels_redis==4.3.0
redis~=6.4.0
Pillow~=11.3.0
bleach~=6.2.0
django-simple-captcha~=0.6.2
eventlet~=0.40.3
django-extensions==4.1
uvicorn[standard]==0.29.0
whitenoise==6.11.0
